+++ src/index.html
@@ -113,6 +113,9 @@
         <div class="btn-loadMore__container">
           <button type="button" class="btn-loadMore__item is-hidden" data-action="load-more">Load more</button>
         </div>
+        <div class="btn-loadMore__container">
+          <button type="button" class="btn-loadMore__item-trend is-hidden" data-action="load-more-trend">Load more</button>
+        </div>
         
         <div id="tui-pagination-container" class="tui-pagination">
           <div id="pagination" class="tui-pagination"></div>
@@ -134,6 +137,7 @@
             </p>
           </div>
         </div>
+      </footer>
     </div>
 
         <div data-modal class="modal_by_students is-hidden">
