+++ src/js/headerLibrary.js
@@ -1,10 +1,18 @@
-import { refs } from './getRefs';
-import MovieApiService from './apiService';
+import Pagination from 'tui-pagination';
+import 'tui-pagination/dist/tui-pagination.css';
 import cardMarkup from '../templates/main-card-markup.hbs';
+import { refs } from './getRefs';
+import { clearGallery, appendMarkup } from './supportFunction';
+import { createPaginationTrend } from './trendPagination';
 import deleteErrorMessage from './deleteErrorMassage';
+import { getTrendMovies } from './trendPagination';
+import FilmApiService from './apiService';
 import LoadMoreBtn from './loadMoreBtn';
 
-const apiService = new MovieApiService();
+const apiService = new FilmApiService();
+const loadMoreButtonTrend = new LoadMoreBtn({
+  selector: '[data-action = "load-more-trend"]',
+});
 const loadMoreButton = new LoadMoreBtn({
   selector: '[data-action = "load-more"]',
 });
@@ -21,10 +29,11 @@ refs.queueBtn.addEventListener('click', (e) => {
 	switchLibraryBtn(refs.queueBtn, refs.watchedBtn)
 })
 
-function openLibrary(evt) {
+function openLibrary(e) {
 	refs.containerPagination.classList.add('is-hidden');
 	loadMoreButton.hide();
-	evt.preventDefault();
+	loadMoreButtonTrend.hide();
+	e.preventDefault();
 	
 	switchLibraryBtn(refs.watchedBtn, refs.queueBtn);
 	deleteErrorMessage();
@@ -33,31 +42,16 @@ function openLibrary(evt) {
 	changeHeader(refs.searchForm, refs.myLibrary, 'second-image', 'first-image');
 };
 
-function openHome(evt) {
-	evt.preventDefault();
+function openHome(e) {
+	e.preventDefault();
 
 	deleteErrorMessage();
 	clearGallery();
-
+	
 	getTrendMovies();
 
 	changeHeader(refs.myLibrary, refs.searchForm, 'first-image', 'second-image');
 };
-
-async function getTrendMovies() {
-	clearGallery();
-
-	try {
-		let movies = await apiService.fetchTrendMovies();
-		appendMarkup(movies.moviesData);
-	} catch (error) {
-		console.log(error);
-	}
-};
-
-function appendMarkup(data) {
-	refs.filmList.insertAdjacentHTML('beforeend', cardMarkup(data))
-}
 
 function switchLibraryBtn(add, remote) {
 	add.classList.add('is-active')
@@ -70,7 +64,3 @@ function changeHeader(addHidden, remoteHidden, addImage, remoteImage) {
 	refs.header.classList.remove(remoteImage);
 	refs.header.classList.add(addImage);
 };
-
-function clearGallery() {
-  refs.filmList.innerHTML = '';
-}
