+++ src/sass/components/_pagination.scss
@@ -15,6 +15,24 @@
 }
 
 .btn-loadMore__item.is-hidden {
+  display: none;
+  opacity: 0;
+  visibility: hidden;
+  pointer-events: none;
+}
+
+.btn-loadMore__item-trend {
+  opacity: 1;
+  visibility: visible;
+  margin: 20px;
+  padding: 10px 40px;
+  border: 1px solid #478fdb;
+  border-radius: 8px;
+  background-color: #4b96e6;
+}
+
+.btn-loadMore__item-trend.is-hidden {
+  display: none;
   opacity: 0;
   visibility: hidden;
   pointer-events: none;
